+++ internal/sessionledger/rand_test.go
@@ -1,0 +1,12 @@
+package sessionledger
+
+import (
+	"crypto/rand"
+	"encoding/hex"
+)
+
+func randomLedgerSuffix() string {
+	var b [6]byte
+	_, _ = rand.Read(b[:])
+	return hex.EncodeToString(b[:])
+}